% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_banks_db.R
\name{build_banks_db}
\alias{build_banks_db}
\title{Downloads quarterly call report data and saves it to local sqlite database.}
\usage{
build_banks_db(db_path)
}
\arguments{
\item{db_path}{Path to sqlite database}
}
\description{
This function is the first one you should use upon downloading this package. It will
download all quarters of data from here, https://www5.fdic.gov/sdi/download_large_list_outside.asp,
store them as zip files in current working directory, unzip them, and write them to a sqlite
database given by db_path.
}
