% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_banks_db.R
\name{query_banks_db}
\alias{query_banks_db}
\title{Helper function to simplify querying database}
\usage{
query_banks_db(
  db_path,
  vars = c(),
  year_start,
  quarter_start,
  year_stop,
  quarter_stop
)
}
\arguments{
\item{db_path}{string}

\item{vars}{character vector of desired data fields}

\item{year_start}{minimum year of data. 1992 through year corresponding to one quarter ago}

\item{quarter_start}{minimum quarter of data. Possible values are 1, 2, 3, or 4}

\item{year_stop}{final year of data. 1992 through year corresponding to one quarter ago}

\item{quarter_stop}{final quarter of data. Possible values are 1, 2, 3, or 4}
}
\value{
tibble
}
\description{
Use this function to easily query the sqlite database that you created. Simpy provide a time range, and specify the
names of the fields that you want to query. This is useful becuase data is spread across multiple tables, and multiple
quarters, and within each quarter, spread across multiple tables. This function allows you to ignore that complexity.
}
\details{
This is intended only to simplify more basic queries, and is by no means intended as the only way to access data
from this database.
}
