% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_banks_db.R
\name{update_banks_db}
\alias{update_banks_db}
\title{Updates bank sqlite database}
\usage{
update_banks_db(db_path)
}
\arguments{
\item{db_path}{Full path to sqlite database, previously created with build_banks_db}
}
\description{
Use this function to update your database of call report data. You should call this approximately quarterly, when
the FDIC provides new data for the previous quarter.
}
