% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_banks.R
\name{write_banks}
\alias{write_banks}
\title{Unzips files, reads in csv contents, writes to sqlite database, and then deletes csv files.}
\usage{
write_banks(db_path, files)
}
\arguments{
\item{db_path}{Path to sqlite database}

\item{files}{Specific files that will be written to sqlite database. Used when updating for new data.}
}
\description{
Unzips files, reads in csv contents, writes to sqlite database, and then deletes csv files.
}
